navbar component and styling
